Move model descriptions to an external file
Description of models are not in the source code anymore. This allows cleaner updates when there are new model versions, as source code does not have to be changed.

Also changed naming convention to only refer to segmentation models as "model", not as "task".

Note: model id in the json file must match the folder name in the .zip file.