Add automatic testing of all models
Click "Reload and test" button on the module widget to test all segmentation models.

All segmentations are executed on GPU and CPU.
Segmentation result is saved as slice sweep and 3D rotation video.
Segmentation times and segment names are recorded.

This can be used for generating a model catalog page.