BUG: Fix segment name containing only modifier (left, right)
When a type was found in a built-in terminology but it did not have modifiers (e.g., scapula left/right) then the segment name was not generated correctly.
Fixed by ensuring that if a type is specified in the MONAIAuto3DSeg terminology then the MONAIAuto3DSeg terminology is used.
This terminology fixes some of the shortcomings of build-in terminologies (such as missing modifiers).

fixes https://github.com/lassoan/SlicerMONAIAuto3DSeg/issues/95